packages/policies/pyproject.toml
[tool.poetry]
name = "noetic-policies"
version = "0.1.0"
description = "Policy specification, parsing, and validation for the Noetic ecosystem"
authors = ["Noetic Team <team@noetic.ai>"]
readme = "README.md"
packages = [{include = "noetic_policies"}]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0"
celpy = "^0.20"
pydantic = "^2.5"
networkx = "^3.6.1"
opentelemetry-api = "^1.22"
opentelemetry-sdk = "^1.22"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^4.1"
hypothesis = "^6.98"
pytest-benchmark = "^4.0"
types-networkx = "^3.6.1"
mypy = "^1.8"
ruff = "^0.1"
black = "^24.0"
pytest-xdist = "^3.5"
pytest-timeout = "^2.2"

[tool.poetry.scripts]
noetic-policies = "noetic_policies.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Mypy configuration (T003)
[tool.mypy]
strict = true
python_version = "3.11"
plugins = ["pydantic.mypy"]
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
check_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

# Ruff configuration (T004)
[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = []

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
"tests/*" = ["S101"]       # Allow assert in tests

# Black configuration (T005)
[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'

# Pytest configuration (T006)
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=noetic_policies",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=80",  # Constitution requirement
]
markers = [
    "fast_mode: Tests for fast validation mode",
    "thorough_mode: Tests for thorough validation mode",
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests",
    "property: Property-based tests",
    "performance: Performance tests",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["noetic_policies"]
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
